requirements.txt
black==22.3.0
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.0
google-api-core==2.7.1
google-api-python-client==2.42.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
importlib-metadata==4.11.3
mypy-extensions==0.4.3
oauth2client==4.1.3
oauthlib==3.2.0
pathspec==0.9.0
pkg-resources==0.0.0
platformdirs==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dateutil==2.8.2
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.2
typing-extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.7.0
